# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/salas": {
            "get": {
                "tags": [
                    "Salas"
                ],
                "summary": "Obtener salas no bloqueadas",
                "description": "Recupera todas las salas cuyo estado no es 'bloqueada'.",
                "operationId": "79600d627c2393b5bd4958c126e993c7",
                "responses": {
                    "200": {
                        "description": "Lista de salas obtenida con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "salas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Sala A"
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "disponible"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron salas no bloqueadas."
                    }
                }
            },
            "post": {
                "tags": [
                    "Salas"
                ],
                "summary": "Crear una nueva sala",
                "description": "Crea una nueva sala con un jugador inicial y establece el estado en 'abierta'.",
                "operationId": "0215164578dcfc36d3f003f7552b3642",
                "requestBody": {
                    "description": "Datos necesarios para crear una sala",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jugador1_id"
                                ],
                                "properties": {
                                    "jugador1_id": {
                                        "description": "ID del jugador que crea la sala",
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sala creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sala": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string",
                                                    "example": "a9b8c7d6-e5f4-3210-b9a8-c7d6e5f43210"
                                                },
                                                "jugador1_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "['abierta', 'bloqueada', 'en_uso', 'terminada']"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada no válidos"
                    }
                }
            }
        },
        "/api/v1/salas/encurso": {
            "get": {
                "tags": [
                    "Salas"
                ],
                "summary": "Obtener salas bloqueadas",
                "description": "Recupera todas las salas cuyo estado es 'bloqueada'.",
                "operationId": "188293ece454467d93d26f9e4f6692aa",
                "responses": {
                    "200": {
                        "description": "Lista de salas bloqueadas obtenida con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "salas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Sala B"
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "bloqueada"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron salas bloqueadas."
                    }
                }
            }
        },
        "/api/v1/salas/{uuid}/unirse": {
            "post": {
                "tags": [
                    "Salas"
                ],
                "summary": "Unirse a una sala",
                "description": "Permite a un jugador unirse a una sala abierta utilizando el UUID de la sala.",
                "operationId": "8b20df64e6b7dcb97613ba2cb17a463c",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID de la sala a la que se desea unir",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "a9b8c7d6-e5f4-3210-b9a8-c7d6e5f43210"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para unirse a la sala",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jugador2_id"
                                ],
                                "properties": {
                                    "jugador2_id": {
                                        "description": "ID del jugador que se une a la sala",
                                        "type": "integer",
                                        "example": 456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugador unido a la sala con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador unido a la sala."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sala no válida o error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El jugador no puede unirse a la sala."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}